feat: Refactor useMessageOption hook for improved modularity and readability
This commit refactors the `useMessageOption` hook to improve its modularity, readability, and maintainability. It extracts chat mode logic and helper functions into separate modules and handlers.

- Extracted chat mode logic into separate modules: `searchChatMode`, `normalChatMode`, `continueChatMode`, and `ragMode`.
- Created helper functions for focus management, input validation, and message saving: `focusTextArea`, `validateBeforeSubmit`, `createSaveMessageOnSuccess`, and `createSaveMessageOnError`.
- Created handlers for message regeneration, editing, and stopping streaming requests: `createRegenerateLastMessage`, `createEditMessage`, and `createStopStreamingRequest`.
- Updated the `useMessageOption` hook to use the extracted modules and handlers.
- Removed unused imports and code.
