+++ src/hooks/chat-modes/normalChatMode.ts
@@ -1,0 +1,324 @@
+import { cleanUrl } from "~/libs/clean-url"
+import {
+  getOllamaURL,
+  systemPromptForNonRagOption
+} from "~/services/ollama"
+import { type ChatHistory, type Message } from "~/store/option"
+import { generateID, getPromptById } from "@/db"
+import { generateHistory } from "@/utils/generate-history"
+import { pageAssistModel } from "@/models"
+import { humanMessageFormatter } from "@/utils/human-message"
+import {
+  isReasoningEnded,
+  isReasoningStarted,
+  mergeReasoningContent
+} from "@/libs/reasoning"
+import { getModelNicknameByID } from "@/db/nickname"
+import { systemPromptFormatter } from "@/utils/system-message"
+
+export const normalChatMode = async (
+  message: string,
+  image: string,
+  isRegenerate: boolean,
+  messages: Message[],
+  history: ChatHistory,
+  signal: AbortSignal,
+  {
+    selectedModel,
+    useOCR,
+    selectedSystemPrompt,
+    currentChatModelSettings,
+    setMessages,
+    saveMessageOnSuccess,
+    saveMessageOnError,
+    setHistory,
+    setIsProcessing,
+    setStreaming,
+    setAbortController,
+    historyId,
+    setHistoryId
+  }: {
+    selectedModel: string
+    useOCR: boolean
+    selectedSystemPrompt: string
+    currentChatModelSettings: any
+    setMessages: (messages: Message[] | ((prev: Message[]) => Message[])) => void
+    saveMessageOnSuccess: (data: any) => Promise<boolean>
+    saveMessageOnError: (data: any) => Promise<boolean>
+    setHistory: (history: ChatHistory) => void
+    setIsProcessing: (value: boolean) => void
+    setStreaming: (value: boolean) => void
+    setAbortController: (controller: AbortController | null) => void
+    historyId: string | null
+    setHistoryId: (id: string) => void
+  }
+) => {
+  console.log("Using normalChatMode")
+  const url = await getOllamaURL()
+  let promptId: string | undefined = selectedSystemPrompt
+  let promptContent: string | undefined = undefined
+
+  if (image.length > 0) {
+    image = `data:image/jpeg;base64,${image.split(",")[1]}`
+  }
+
+  const ollama = await pageAssistModel({
+    model: selectedModel!,
+    baseUrl: cleanUrl(url)
+  })
+
+  let newMessage: Message[] = []
+  let generateMessageId = generateID()
+  const modelInfo = await getModelNicknameByID(selectedModel)
+
+  if (!isRegenerate) {
+    newMessage = [
+      ...messages,
+      {
+        isBot: false,
+        name: "You",
+        message,
+        sources: [],
+        images: [image],
+        modelImage: modelInfo?.model_avatar,
+        modelName: modelInfo?.model_name || selectedModel
+      },
+      {
+        isBot: true,
+        name: selectedModel,
+        message: "▋",
+        sources: [],
+        id: generateMessageId,
+        modelImage: modelInfo?.model_avatar,
+        modelName: modelInfo?.model_name || selectedModel
+      }
+    ]
+  } else {
+    newMessage = [
+      ...messages,
+      {
+        isBot: true,
+        name: selectedModel,
+        message: "▋",
+        sources: [],
+        id: generateMessageId,
+        modelImage: modelInfo?.model_avatar,
+        modelName: modelInfo?.model_name || selectedModel
+      }
+    ]
+  }
+  setMessages(newMessage)
+  let fullText = ""
+  let contentToSave = ""
+  let timetaken = 0
+
+  try {
+    const prompt = await systemPromptForNonRagOption()
+    const selectedPrompt = await getPromptById(selectedSystemPrompt)
+
+    let humanMessage = await humanMessageFormatter({
+      content: [
+        {
+          text: message,
+          type: "text"
+        }
+      ],
+      model: selectedModel,
+      useOCR: useOCR
+    })
+    if (image.length > 0) {
+      humanMessage = await humanMessageFormatter({
+        content: [
+          {
+            text: message,
+            type: "text"
+          },
+          {
+            image_url: image,
+            type: "image_url"
+          }
+        ],
+        model: selectedModel,
+        useOCR: useOCR
+      })
+    }
+
+    const applicationChatHistory = generateHistory(history, selectedModel)
+
+    if (prompt && !selectedPrompt) {
+      applicationChatHistory.unshift(
+        await systemPromptFormatter({
+          content: prompt
+        })
+      )
+    }
+
+    const isTempSystemprompt =
+      currentChatModelSettings.systemPrompt &&
+      currentChatModelSettings.systemPrompt?.trim().length > 0
+
+    if (!isTempSystemprompt && selectedPrompt) {
+      applicationChatHistory.unshift(
+        await systemPromptFormatter({
+          content: selectedPrompt.content
+        })
+      )
+      promptContent = selectedPrompt.content
+    }
+
+    if (isTempSystemprompt) {
+      applicationChatHistory.unshift(
+        await systemPromptFormatter({
+          content: currentChatModelSettings.systemPrompt
+        })
+      )
+      promptContent = currentChatModelSettings.systemPrompt
+    }
+
+    let generationInfo: any | undefined = undefined
+
+    const chunks = await ollama.stream(
+      [...applicationChatHistory, humanMessage],
+      {
+        signal: signal,
+        callbacks: [
+          {
+            handleLLMEnd(output: any): any {
+              try {
+                generationInfo = output?.generations?.[0][0]?.generationInfo
+              } catch (e) {
+                console.error("handleLLMEnd error", e)
+              }
+            }
+          }
+        ]
+      }
+    )
+
+    let count = 0
+    let reasoningStartTime: Date | null = null
+    let reasoningEndTime: Date | null = null
+    let apiReasoning: boolean = false
+
+    for await (const chunk of chunks) { 
+      if (chunk?.additional_kwargs?.reasoning_content) {
+        const reasoningContent = mergeReasoningContent(
+          fullText,
+          chunk?.additional_kwargs?.reasoning_content || ""
+        )
+        contentToSave = reasoningContent
+        fullText = reasoningContent
+        apiReasoning = true
+      } else {
+        if (apiReasoning) {
+          fullText += "</think>"
+          contentToSave += "</think>"
+          apiReasoning = false
+        }
+      }
+
+      contentToSave += chunk?.content
+      fullText += chunk?.content
+
+      if (isReasoningStarted(fullText) && !reasoningStartTime) {
+        reasoningStartTime = new Date()
+      }
+
+      if (
+        reasoningStartTime &&
+        !reasoningEndTime &&
+        isReasoningEnded(fullText)
+      ) {
+        reasoningEndTime = new Date()
+        const reasoningTime =
+          reasoningEndTime.getTime() - reasoningStartTime.getTime()
+        timetaken = reasoningTime
+      }
+
+      if (count === 0) {
+        setIsProcessing(true)
+      }
+      setMessages((prev) => {
+        return prev.map((message) => {
+          if (message.id === generateMessageId) {
+            return {
+              ...message,
+              message: fullText + "▋",
+              reasoning_time_taken: timetaken
+            }
+          }
+          return message
+        })
+      })
+      count++
+    }
+
+    setMessages((prev) => {
+      return prev.map((message) => {
+        if (message.id === generateMessageId) {
+          return {
+            ...message,
+            message: fullText,
+            generationInfo,
+            reasoning_time_taken: timetaken
+          }
+        }
+        return message
+      })
+    })
+
+    setHistory([
+      ...history,
+      {
+        role: "user",
+        content: message,
+        image
+      },
+      {
+        role: "assistant",
+        content: fullText
+      }
+    ])
+
+    await saveMessageOnSuccess({
+      historyId,
+      setHistoryId,
+      isRegenerate,
+      selectedModel: selectedModel,
+      message,
+      image,
+      fullText,
+      source: [],
+      generationInfo,
+      prompt_content: promptContent,
+      prompt_id: promptId,
+      reasoning_time_taken: timetaken
+    })
+
+    setIsProcessing(false)
+    setStreaming(false)
+  } catch (e) {
+    const errorSave = await saveMessageOnError({
+      e,
+      botMessage: fullText,
+      history,
+      historyId,
+      image,
+      selectedModel,
+      setHistory,
+      setHistoryId,
+      userMessage: message,
+      isRegenerating: isRegenerate,
+      prompt_content: promptContent,
+      prompt_id: promptId
+    })
+
+    if (!errorSave) {
+      throw e // Re-throw to be handled by the calling function
+    }
+    setIsProcessing(false)
+    setStreaming(false)
+  } finally {
+    setAbortController(null)
+  }
+}
